% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROACT-package.R
\docType{package}
\name{PROACT-package}
\alias{PROACT}
\alias{PROACT-package}
\title{PROACT: Prior-Regularized Optimization for Analysis of Co-occurring TFs}
\description{
PROACT provides functions for constructing TF prior matrices, preparing binary
peak-by-TF matrices, inferring candidate cooperative TF associations with
priority LASSO, characterizing spatial motif grammar, and visualizing TF-TF networks.
}
\details{
The package bundles its motif annotation as \code{inst/extdata/metadata.tsv}.
PPI evidence files are not bundled and must be supplied by the user.
}
\section{Main functions}{

\itemize{
\item \code{\link[=build_prior]{build_prior()}} constructs the TF prior matrix.
\item \code{\link[=prepare_peak_tf_matrix]{prepare_peak_tf_matrix()}} prepares the binary peak-by-TF matrix.
\item \code{\link[=run_PROACT]{run_PROACT()}} performs prior-regularized cooperative TF inference.
\item \code{\link[=find_grammar]{find_grammar()}}, \code{\link[=analyze_tf_grammar]{analyze_tf_grammar()}}, \code{\link[=summarize_tf_grammar]{summarize_tf_grammar()}},
and \code{\link[=plot_tf_grammar]{plot_tf_grammar()}} characterize TF-pair spatial grammar.
\item \code{\link[=plot_PROACT_network]{plot_PROACT_network()}} visualizes inferred TF association networks.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hdm2020/PROACT}
  \item Report bugs at \url{https://github.com/hdm2020/PROACT/issues}
}

}
\author{
\strong{Maintainer}: Dongmei Han \email{handongmei@ucas.ac.cn}

Authors:
\itemize{
  \item Dongmei Han \email{handongmei@ucas.ac.cn}
}

}
\keyword{internal}
